requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
numpy==1.19.5
tensorflow-cpu==2.6.0
keras==2.6.0
gunicorn==19.7.1
pillow==5.0.0
googledrivedownloader==0.4
#tensorflow==2.2.1
#tensorflow==2.6.0
# absl-py==0.1.13
# astor==0.6.2
# bleach==1.5.0
# click==6.7
# Flask==0.12.2
# gast==0.2.0
# grpcio==1.10.0
# gunicorn==19.7.1
# h5py==2.7.1
# html5lib==0.9999999
# itsdangerous==0.24
# Jinja2==2.10
# Keras==2.1.5
# Markdown==2.6.11
# # MarkupSafe==1.0
# numpy==1.14.2
# opencv-python==3.4.0.12
# pillow==5.0.0
# protobuf==3.5.2.post1
# PyYAML==3.12
# scipy==1.0.1
# six==1.11.0
# tensorboard==1.6.0
# tensorflow==1.6.0
# termcolor==1.1.0
# Werkzeug==0.14.1
